rag/pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "RAG application with PDF and document processing"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    "pypdf>=5.0.0",
    "python-docx>=1.1.2",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "chromadb>=0.5.0",
    "openai>=1.51.0",
    "sentence-transformers>=3.1.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "huggingface-hub>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["app.py", "rag_engine.py"]
